Update find numebrs
Changed a lot on this. Added scrape feature to find numbers that are not links, added formatting of phone numbers to remove special characters and only send numbers. Added custom prefix function to allow setting prefix in options. Changed how phone numbers are stored. They are now assigned to the link created as it's class attribute so that there can be more than one on a page and it will not always send the last one stored in tel variable.